CHANGES: - Auto affinity now supports NUMA nodes (CCDs)

+++ core/data/proxy/nodeproxy.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import json
 import os
 
 from core.data.node import Node, UploadStatus, SortOrder
@@ -441,21 +442,26 @@ class NodeProxy(Node):
 
     @override
     @async_cache
-    async def get_cpu_info(self, used: bool = True) -> bytes | int:
+    async def get_cpu_info(self, used: bool = True, export: bool = False) -> bytes | dict | int:
         timeout = 60 if not self.slow_system else 120
         data = await self.bus.send_to_node_sync({
             "command": "rpc",
             "object": "Node",
             "method": "get_cpu_info",
             "params": {
-                "used": used
+                "used": used,
+                "export": export
             }
         }, timeout=timeout, node=self.name)
+        if isinstance(data, dict):
+            return data
         async with self.apool.connection() as conn:
             cursor = await conn.execute("SELECT data FROM files WHERE id = %s", (data, ), binary=True)
-            image = (await cursor.fetchone())[0]
+            result = (await cursor.fetchone())[0]
             await conn.execute("DELETE FROM files WHERE id = %s", (data, ))
-        return image
+        if export:
+            return json.loads(result)
+        return result
 
     @override
     async def info(self) -> dict:
